requirements.txt
python-telegram-bot
geopy
beautifulsoup4
requests
lxml
psycopg2
pymongo
dnspython